+++ src/main/php/lang/ast/emit/PHP56.class.php
@@ -1,16 +1,17 @@
 <?php namespace lang\ast\emit;
 
+use lang\ast\Emitter;
 use lang\ast\nodes\Value;
 
 /**
  * PHP 5.6 syntax
  *
- * @see  https://wiki.php.net/rfc/pow-operator
- * @see  https://wiki.php.net/rfc/variadics
- * @see  https://wiki.php.net/rfc/argument_unpacking
- * @see  https://wiki.php.net/rfc/use_function
+ * @see  https://wiki.php.net/rfc#php_56
  */
-class PHP56 extends \lang\ast\Emitter {
+class PHP56 extends Emitter {
+  use OmitPropertyTypes, OmitReturnTypes, OmitConstModifiers;
+  use RewriteLambdaExpressions;
+
   protected $unsupported= [
     'object'   => 72,
     'void'     => 71,
@@ -89,9 +90,6 @@ class PHP56 extends \lang\ast\Emitter {
     'parent'       => true
   ];
 
-  protected function returnType($name) {
-    return null;
-  }
 
   protected function emitLiteral($literal) {
     if ('"' === $literal{0}) {
@@ -121,12 +119,6 @@ class PHP56 extends \lang\ast\Emitter {
     $this->out->write('}');
   }
 
-  protected function emitConst($const) {
-    $this->out->write('const '.$const->name.'=');
-    $this->emit($const->expression);
-    $this->out->write(';');
-  }
-
   protected function emitBinary($binary) {
     if ('??' === $binary->operator) {
       $this->out->write('isset(');
@@ -145,6 +137,21 @@ class PHP56 extends \lang\ast\Emitter {
       $this->out->write(') ? -1 : ('.$l.' == '.$r.' ? 0 : 1)');
     } else {
       parent::emitBinary($binary);
+    }
+  }
+
+  protected function emitAssignment($assignment) {
+    if ('??=' === $assignment->operator) {
+      $this->out->write('isset(');
+      $this->emitAssign($assignment->variable);
+      $this->out->write(') ||');
+      $this->emit($assignment->variable);
+      $this->out->write('=');
+      $this->emit($assignment->expression);
+    } else {
+      $this->emitAssign($assignment->variable);
+      $this->out->write($assignment->operator);
+      $this->emit($assignment->expression);
     }
   }