+++ src/test/php/lang/ast/unittest/parse/TypesTest.class.php
@@ -1,6 +1,6 @@
 <?php namespace lang\ast\unittest\parse;
 
-use lang\ast\Error;
+use lang\ast\Errors;
 
 class TypesTest extends ParseTest {
 
@@ -116,17 +116,17 @@ class TypesTest extends ParseTest {
     );
   }
 
-  #[@test, @expect(class= Error::class, withMessage= 'Cannot redeclare method b()')]
+  #[@test, @expect(class= Errors::class, withMessage= 'Cannot redeclare method b()')]
   public function cannot_redeclare_method() {
     iterator_to_array($this->parse('class A { public function b() { } public function b() { }}'));
   }
 
-  #[@test, @expect(class= Error::class, withMessage= 'Cannot redeclare property $b')]
+  #[@test, @expect(class= Errors::class, withMessage= 'Cannot redeclare property $b')]
   public function cannot_redeclare_property() {
     iterator_to_array($this->parse('class A { public $b; private $b; }'));
   }
 
-  #[@test, @expect(class= Error::class, withMessage= 'Cannot redeclare constant B')]
+  #[@test, @expect(class= Errors::class, withMessage= 'Cannot redeclare constant B')]
   public function cannot_redeclare_constant() {
     iterator_to_array($this->parse('class A { const B = 1; const B = 3; }'));
   }